Added CHKROOTKIT, update size list
Now asks to install chkrootkit as well. Checks with the bum varible if it was downloaded. A limitation to this system is if somebody selects N even when it is installed, it will not run. CHKROOTKIT goes through a grep to only show if files are infected. Also, dir sizes output to a file called `dirsizes.txt' using "tee"